google matrix api

+++ src/main/java/com/foodie/foodieapp/utils/GeoUtils.java
@@ -15,6 +15,8 @@ public class GeoUtils {
 
         double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
 
-        return EARTH_RADIUS_KM * c;
+        return (int) (EARTH_RADIUS_KM * c);
+
+        
     }
 }
